0419 放視最終檔

+++ Noir/Assets/Script/Player/Controller/UI/UI_HP.cs
@@ -96,7 +96,7 @@ public class UI_HP : MonoBehaviour
         darkBarCanMove = true;
     }
 
-    public void ConsumeHp(int consumeHP)
+    public void ConsumeHp(float consumeHP)
     {
         HP -= consumeHP;
         DarkBarStartControl();
